feat: moves core_models into isolated module
