+++ src/aind_metadata_mapper/bergamo/models.py
@@ -3,7 +3,7 @@
 from decimal import Decimal
 from typing import List, Literal, Optional
 
-from aind_metadata_mapper.core import BaseJobSettings
+from aind_metadata_mapper.core_models import BaseJobSettings
 
 
 class JobSettings(BaseJobSettings):
