+++ src/aind_metadata_mapper/bruker/session.py
@@ -32,7 +32,8 @@ from aind_data_schema_models.units import TimeUnit
 from bruker2nifti._metadata import BrukerMetadata
 
 from aind_metadata_mapper.bruker.models import JobSettings
-from aind_metadata_mapper.core import GenericEtl, JobResponse
+from aind_metadata_mapper.core import GenericEtl
+from aind_metadata_mapper.core_models import JobResponse
 
 DATETIME_FORMAT = "%H:%M:%S %d %b %Y"
 LENGTH_FORMAT = "%Hh%Mm%Ss%fms"
