+++ src/aind_metadata_mapper/fip/models.py
@@ -3,7 +3,7 @@
 from datetime import datetime
 from typing import List, Literal
 
-from aind_metadata_mapper.core import BaseJobSettings
+from aind_metadata_mapper.core_models import BaseJobSettings
 
 
 class JobSettings(BaseJobSettings):
