+++ src/aind_metadata_mapper/mesoscope/models.py
@@ -6,7 +6,7 @@ from typing import List, Literal, Union
 
 from pydantic import Field
 
-from aind_metadata_mapper.core import BaseJobSettings
+from aind_metadata_mapper.core_models import BaseJobSettings
 
 
 class JobSettings(BaseJobSettings):
