+++ src/aind_metadata_mapper/u19/procedures.py
@@ -16,7 +16,8 @@ from aind_data_schema.core.procedures import (
 )
 from aind_data_schema_models.organizations import Organization
 
-from aind_metadata_mapper.core import GenericEtl, JobResponse
+from aind_metadata_mapper.core import GenericEtl
+from aind_metadata_mapper.core_models import JobResponse
 from aind_metadata_mapper.u19.models import JobSettings
 from aind_metadata_mapper.u19.utils import construct_new_model
